Run linter,format & refactor code
