+++ nft-module/package.json
@@ -21,21 +21,22 @@
   "dependencies": {
     "@kodadot1/uniquery": "^0.6.0-rc.0",
     "@nestjs/common": "^10.3.2",
+    "@nestjs/config": "^3.2.3",
     "@nestjs/core": "^10.3.2",
     "@nestjs/platform-express": "^10.3.2",
     "cors": "^2.8.5",
     "reflect-metadata": "^0.2.1",
-    "rxjs": "^7.8.1"
+    "rxjs": "^7.8.1",
+    "@polkadot/api": "^12.1.1",
+    "@polkadot/api-base": "^12.1.1",
+    "@polkadot/apps-config": "^0.141.1",
+    "@polkadot/types": "^12.1.1",
+    "@polkadot/util": "^12.6.2"
   },
   "devDependencies": {
     "@nestjs/cli": "^10.3.1",
     "@nestjs/schematics": "^10.1.0",
     "@nestjs/testing": "^10.3.2",
-    "@polkadot/api": "^12.1.1",
-    "@polkadot/api-base": "^12.1.1",
-    "@polkadot/apps-config": "^0.141.1",
-    "@polkadot/types": "^12.1.1",
-    "@polkadot/util": "^12.6.2",
     "@swc/cli": "^0.3.9",
     "@swc/core": "^1.4.0",
     "@types/express": "^4.17.21",
@@ -74,6 +75,5 @@
     ],
     "coverageDirectory": "../coverage",
     "testEnvironment": "node"
-  },
-  "packageManager": "pnpm@9.4.0+sha512.f549b8a52c9d2b8536762f99c0722205efc5af913e77835dbccc3b0b0b2ca9e7dc8022b78062c17291c48e88749c70ce88eb5a74f1fa8c4bf5e18bb46c8bd83a"
+  }
 }
