+++ nft-module/src/app.module.ts
@@ -1,13 +1,20 @@
-import { Module } from '@nestjs/common';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
-import { CollectionModule } from './collectionCreator/collection.module';
-import { MetadataModule } from './nftMetadata/metadata.module';
-import { NFTModule } from './nftMinter/nft.module';
-import { SwapModule } from './ownershipSwap/swap.module';
+import { Module } from "@nestjs/common";
+import { AppController } from "./app.controller";
+import { AppService } from "./app.service";
+import { CollectionModule } from "./modules/collectionCreator/collection.module";
+import { MetadataModule } from "./modules/nftMetadata/metadata.module";
+import { NFTModule } from "./modules/nftMinter/nft.module";
+import { SwapModule } from "./modules/ownershipSwap/swap.module";
+import { ConfigModule } from "@nestjs/config";
 
 @Module({
-  imports: [NFTModule,CollectionModule,MetadataModule, SwapModule],
+  imports: [
+    ConfigModule.forRoot({ isGlobal: true }),
+    NFTModule,
+    CollectionModule,
+    MetadataModule,
+    SwapModule,
+  ],
   controllers: [AppController],
   providers: [AppService],
 })
