+++ nft-module/src/common/config/app-config.ts
@@ -1,0 +1,5 @@
+export interface AppConfig {
+  WSS_ENDPOINT: string;
+  BACKEND_URL: string;
+  PORT: string;
+}
