+++ nft-module/src/common/config/index.ts
@@ -1,0 +1,1 @@
+export * from "./app-config";
