+++ nft-module/src/main.ts
@@ -1,16 +1,18 @@
-import { NestFactory } from '@nestjs/core';
-import { AppModule } from './app.module';
+import { NestFactory } from "@nestjs/core";
+import { AppModule } from "./app.module";
+import { ConfigService } from "@nestjs/config";
+import { AppConfig } from "@common/config";
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
-  
+  const config = app.get<ConfigService<AppConfig>>(ConfigService);
+
   // Enable CORS
-  /*app.enableCors({
-    origin: 'http://localhost:5173', // Replace with your frontend URL
-    methods: 'GET,HEAD,PUT,PATCH,POST,DELETE,OPTIONS',
+  app.enableCors({
+    origin: config.get("BACKEND_URL"),
+    methods: "GET,HEAD,PUT,PATCH,POST,DELETE,OPTIONS",
     credentials: true,
-  }); 
-  */
+  });
   await app.listen(3000);
 }
 bootstrap();
