+++ nft-module/src/modules/collectionCreator/collection.module.ts
@@ -1,0 +1,10 @@
+import { Module } from "@nestjs/common";
+import { CollectionController } from "./collection.controller";
+import { collectionCreator } from "./collection.service";
+
+@Module({
+  imports: [],
+  controllers: [CollectionController],
+  providers: [collectionCreator],
+})
+export class CollectionModule {}
