+++ nft-module/src/modules/nftMetadata/dto/AccountDto.ts
@@ -1,4 +1,4 @@
-import { IsNotEmpty } from 'class-validator';
+import { IsNotEmpty } from "class-validator";
 
 export class AccountDto {
   @IsNotEmpty()
