+++ nft-module/src/modules/ownershipSwap/dto/SwapDto.ts
@@ -1,0 +1,10 @@
+import { IsNotEmpty } from "class-validator";
+
+export class SwapDto {
+  @IsNotEmpty()
+  assetID: string;
+  @IsNotEmpty()
+  collectionID: string;
+  @IsNotEmpty()
+  address: string;
+}
