--- nft-module/src/nftMetadata/metadata.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import { MetadataController } from './metadata.controller';
-import { nftModuleService } from './metadata.service';
-
-@Module({
-  imports: [],
-  controllers: [MetadataController],
-  providers: [nftModuleService],
-})
-export class MetadataModule {}
