--- nft-module/src/nftMinter/dto/NftDto.ts
@@ -1,19 +1,0 @@
-import { IsNotEmpty } from 'class-validator';
-
-export class Metadata {
-    @IsNotEmpty()
-    name: string;
-    @IsNotEmpty()
-    description: string;
-    @IsNotEmpty()
-    ipfs: string;
-}
-
-export class NftDto {
-    @IsNotEmpty()
-    metadata: Metadata;
-    @IsNotEmpty()
-    address: string;
-    @IsNotEmpty()
-    collectionID: number;
-}
