--- nft-module/src/nftMinter/nft.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import { NftController } from './nft.controller';
-import { nftCreator } from './nft.service';
-
-@Module({
-  imports: [],
-  controllers: [NftController],
-  providers: [nftCreator],
-})
-export class NFTModule {}
