--- nft-module/src/ownershipSwap/dto/SwapDto.ts
@@ -1,10 +1,0 @@
-import { IsNotEmpty } from 'class-validator';
-
-export class SwapDto {
-    @IsNotEmpty()
-    assetID: string;
-    @IsNotEmpty()
-    collectionID: string;
-    @IsNotEmpty()
-    address: string;
-}
